backend/requirements.txt
playwright==1.40.0
requests==2.31.0
pymupdf==1.23.0
pdfplumber==0.10.3
pytesseract==0.3.10
psycopg2-binary==2.9.9
tiktoken==0.5.2
openai>=1.40.0
aiohttp==3.9.1